requirements.txt
black==19.3b0
bunq-sdk==1.10.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0